File from NY state with data from 2001
https://data.ny.gov/Economic-Development/Current-Employment-Statistics-Beginning-1990/6k74-dgkb 

Last updated: November 17, 2023

Current Employment by Industry (CES) data reflect jobs by "place of work." It does not include the self-employed, unpaid family workers, and private household employees. Jobs located in the county or the metropolitan area that pay wages and salaries are counted although workers may live outside the area. Jobs are counted regardless of the number of hours worked. Individuals who hold more than one job (i.e. multiple job holders) may be counted more than once. The employment figure is an estimate of the number of jobs in the area (regardless of the place of residence of the workers) rather than a count of jobs held by the residents of the area.